+++ lib/rspec/core/command_line.rb
@@ -12,44 +12,22 @@ module RSpec
       end
 
       def run(err, out)
-        @options.configure(@configuration)
         @configuration.error_stream = err
         @configuration.output_stream ||= out
+        @options.configure(@configuration)
         @configuration.load_spec_files
         @configuration.configure_mock_framework
         @world.announce_inclusion_filter
         @world.announce_exclusion_filter
 
-        @configuration.reporter.report(example_count) do |reporter|
+        @configuration.reporter.report(@world.example_count) do |reporter|
           begin
             @configuration.run_hook(:before, :suite)
-            example_groups.run_examples(reporter)
+            @world.example_groups.map {|g| g.run(reporter)}.all?
           ensure
             @configuration.run_hook(:after, :suite)
           end
         end
-
-        example_groups.success?
-      end
-
-    private
-
-      def example_count
-        @world.example_count
-      end
-
-      module ExampleGroups
-        def run_examples(reporter)
-          @success = self.inject(true) {|success, group| success &= group.run(reporter)}
-        end
-
-        def success?
-          @success ||= false
-        end
-      end
-
-      def example_groups
-        @world.example_groups.extend(ExampleGroups)
       end
     end
   end